Refactor Lang.Expr module
