+++ hschain-utxo-lang/hschain-utxo-lang.cabal
@@ -55,6 +55,8 @@ library
       Hschain.Utxo.Lang.Expr
       Hschain.Utxo.Lang.Const
       Hschain.Utxo.Lang.Infer
+      Hschain.Utxo.Lang.Module
+      Hschain.Utxo.Lang.UserType
       Hschain.Utxo.Lang.Lib.Base
       Hschain.Utxo.Lang.Monad
       Hschain.Utxo.Lang.Parser.Hask
