+++ hschain-utxo-lang/src/Hschain/Utxo/Lang.hs
@@ -11,7 +11,9 @@ import Hschain.Utxo.Lang.Sigma      as X
 import Hschain.Utxo.Lang.Expr       as X
 import Hschain.Utxo.Lang.Exec       as X
 import Hschain.Utxo.Lang.Pretty     as X
+import Hschain.Utxo.Lang.Module     as X
 import Hschain.Utxo.Lang.Types      as X
+import Hschain.Utxo.Lang.UserType   as X
 import Hschain.Utxo.Lang.Parser.Quoter as X
 import Hschain.Utxo.Lang.Utils.ByteString as X
