+++ hschain-utxo-lang/src/Hschain/Utxo/Lang/Desugar/Case.hs
@@ -15,6 +15,8 @@ import Data.Maybe
 import Data.Sequence (Seq)
 
 import Hschain.Utxo.Lang.Expr
+import Hschain.Utxo.Lang.Module
+import Hschain.Utxo.Lang.UserType
 import Hschain.Utxo.Lang.Core.Types (Prim(..))
 import Hschain.Utxo.Lang.Exec.Subst
 import Hschain.Utxo.Lang.Monad
