+++ hschain-utxo-lang/src/Hschain/Utxo/Lang/Desugar/Records.hs
@@ -11,6 +11,7 @@ import Control.Arrow (first)
 
 import Hschain.Utxo.Lang.Expr
 import Hschain.Utxo.Lang.Monad
+import Hschain.Utxo.Lang.UserType
 
 import Data.Fix
 import Data.Maybe
