+++ hschain-utxo-lang/src/Hschain/Utxo/Lang/Exec/Module.hs
@@ -28,6 +28,8 @@ import Hschain.Utxo.Lang.Const (reservedNames)
 import Hschain.Utxo.Lang.Desugar
 import Hschain.Utxo.Lang.Error
 import Hschain.Utxo.Lang.Expr
+import Hschain.Utxo.Lang.Module
+import Hschain.Utxo.Lang.UserType
 import Hschain.Utxo.Lang.Infer
 import Hschain.Utxo.Lang.Monad
 import Hschain.Utxo.Lang.Lib.Base (baseNames, baseLibTypes)
