+++ hschain-utxo-lang/src/Hschain/Utxo/Lang/Expr.hs
@@ -1,12 +1,46 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 -- | This module defines AST for the language
-module Hschain.Utxo.Lang.Expr where
+module Hschain.Utxo.Lang.Expr(
+    Loc
+  , noLoc
+  , VarName(..)
+  , ConsName(..)
+  , consToVarName
+  , varToConsName
+  , Lang
+  , E(..)
+  , Type
+  , TypeError
+  , Signature
+  , CaseExpr(..)
+  , Binds(..)
+  , Bind(..)
+  , Alt(..)
+  , Pat(..)
+  , patNames
+  , Rhs(..)
+  , Guard(..)
+  , QuoteType(..)
+  , TypeContext
+  , sortBinds
+  , freeVars
+  , freeVarsPat
+  , freeVarsAlt
+  , bindNames
+  , bindNamesLhs
+  , bindAlts
+  , getBindsNames
+  , secretVar
+  , BoolExprResult(..)
+  , mapDeclsM
+  , fromParserLoc
+  , emptyTypeContext
+  , ToLang(..)
+) where
 
 import Hex.Common.Aeson
-import Hex.Common.Text
 
 import Control.Applicative
-import Control.Monad.State.Strict
 
 import Data.Aeson
 import Data.ByteString (ByteString)
@@ -17,7 +51,6 @@ import Data.Foldable
 import Data.Graph
 import Data.Int
 import Data.Map.Strict (Map)
-import Data.Maybe
 import Data.String
 import Data.Set (Set)
 import Data.Text (Text)
@@ -31,19 +64,14 @@ import Text.Show.Deriving
 
 import HSChain.Crypto.Classes (ByteRepr(..))
 import Hschain.Utxo.Lang.Sigma
-import Hschain.Utxo.Lang.Core.Types         (TypeCore(..), argsTuple, Name)
 import Hschain.Utxo.Lang.Types              (Script(..))
-import Hschain.Utxo.Lang.Core.Compile.Expr  (PrimOp(..))
 import Hschain.Utxo.Lang.Core.Types         (Prim(..))
 import qualified Type.Check.HM as H
 import qualified Language.Haskell.Exts.SrcLoc as Hask
 
-import qualified Data.Map.Strict as M
 import qualified Data.List as L
 import qualified Data.Set as Set
 import qualified Data.Vector as V
-
-import qualified Hschain.Utxo.Lang.Const as Const
 
 type Loc = Hask.SrcSpanInfo
 type Type = H.Type Loc Text
@@ -59,251 +87,6 @@ instance H.DefLoc Hask.SrcSpanInfo where
 -- | Unknown source code location.
 noLoc :: Loc
 noLoc = Hask.noSrcSpan
-
--- | Context of user-defined types
-data UserTypeCtx = UserTypeCtx
-  { userTypeCtx'types      :: Map VarName  UserType          -- ^ User-defined types
-  , userTypeCtx'constrs    :: Map ConsName ConsInfo          -- ^ Map from constructor names to it's low-level data, for further compilation
-  , userTypeCtx'recConstrs :: Map ConsName RecordFieldOrder  -- ^ Order of fields for records
-  , userTypeCtx'recFields  :: Map Text     (ConsName, RecordFieldOrder)  -- ^ Maps record single field to the full lists of fields
-  , userTypeCtx'core       :: UserCoreTypeCtx
-  }
-  deriving (Generic, Data, Typeable)
-  deriving (Semigroup, Monoid) via GenericSemigroupMonoid UserTypeCtx
-
-data UserCoreTypeCtx = UserCoreTypeCtx
-  { userCoreTypeCtx'types   :: Map Name (H.Signature () Name)
-  , userCoreTypeCtx'constrs :: Map Name CoreConsDef
-  }
-  deriving (Show, Eq, Generic, Data, Typeable)
-  deriving (Semigroup, Monoid) via GenericSemigroupMonoid UserCoreTypeCtx
-
-userCoreTypeMap :: Map Name UserType -> UserCoreTypeCtx
-userCoreTypeMap ts = execState (mapM_ go $ M.toList ts) (UserCoreTypeCtx mempty mempty)
-  where
-    go :: (Name, UserType) -> State UserCoreTypeCtx ()
-    go (name, def)
-      | isBaseType name = return ()
-      | otherwise       = do
-          st <- get
-          case M.lookup name $ userCoreTypeCtx'types st of
-            Just _   -> return ()
-            Nothing  -> do
-              t <- convertUserType def
-              modify' $ \env -> env { userCoreTypeCtx'types = M.insert name t $ userCoreTypeCtx'types env }
-
-    isBaseType x = Set.member x baseTypes
-    baseTypes = Set.fromList ["Maybe"]
-
-    convertUserType UserType{..} = fmap toSig $ do
-      mapM_ addCons constrs
-      case constrs of
-        []         -> pure unitT
-        [(_, def)] -> defToTuple def
-        _          -> fmap (H.conT () ("Sum" <> showt (length constrs))) $ mapM (defToTuple . snd) constrs
-      where
-        toSig t = foldr (\v ty -> H.forAllT () (varName'name v) ty) (H.monoT t) userType'args
-
-        constrs = M.toList userType'cases
-        constrOrderMap = M.fromList $ zipWith (\n (name, _) -> (name, n)) [0..] constrs
-
-        getOrder n = M.lookup n constrOrderMap
-
-        defToTuple x = case V.toList $ getConsTypes x of
-          []  -> pure unitT
-          [t] -> convertType $ eraseLoc t
-          tys -> fmap tupleT $ mapM (convertType . eraseLoc) tys
-
-        getSumTs
-          | length constrs < 2 = pure Nothing
-          | otherwise          = fmap (Just . V.fromList) $ mapM (toSumArg . snd) constrs
-
-        toSumArg :: ConsDef -> State UserCoreTypeCtx (H.Type () Name)
-        toSumArg def
-          | arity == 0 = pure unitT
-          | arity == 1 = convertType $ eraseLoc $ V.head tys
-          | otherwise  = fmap (tupleT . V.toList) $ mapM (convertType . eraseLoc) tys
-          where
-            arity = V.length tys
-            tys = getConsTypes def
-
-        toTup = mapM (convertType . eraseLoc) . getConsTypes
-
-        addCons (name, def) = do
-          sumTs <- getSumTs
-          tupleTs <- toTup def
-          let coreDef = CoreConsDef
-                { coreConsDef'sum   = liftA2 (,) (getOrder name) sumTs
-                , coreConsDef'tuple = tupleTs
-                }
-          modify' $ \st -> st { userCoreTypeCtx'constrs = M.insert (consName'name name) coreDef $ userCoreTypeCtx'constrs st }
-
-    convertType :: H.Type () Name -> State UserCoreTypeCtx (H.Type () Name)
-    convertType (H.Type x) = fmap H.Type $ foldFixM tfm x
-      where
-        tfm = \case
-          H.ConT loc con args | isUserType con -> do
-            st <- get
-            case M.lookup con $ userCoreTypeCtx'types st of
-              Just t  -> pure $ H.unType $ H.closeSignature (fmap H.Type args) t
-              Nothing -> case M.lookup con ts of
-                           Just defn -> do
-                             t <- convertUserType defn
-                             modify' $ \env -> env { userCoreTypeCtx'types = M.insert con t $ userCoreTypeCtx'types env }
-                             return (H.unType $ H.closeSignature (fmap H.Type args) t)
-                           Nothing -> pure $ Fix $ H.ConT loc con args
-          other               -> pure $ Fix $ other
-
-    isUserType x = Set.member x allTypeNames
-    allTypeNames = Set.fromList $ M.keys ts
-
-    eraseLoc = H.setLoc ()
-
-recFieldExecCtx :: UserTypeCtx -> ExecCtx
-recFieldExecCtx UserTypeCtx{..} =
-  ExecCtx $ M.mapKeys (VarName noLoc) $ M.mapWithKey recSelectorExpr userTypeCtx'recFields
-
-recSelectorExpr :: Text -> (ConsName, RecordFieldOrder) -> Lang
-recSelectorExpr field (cons, RecordFieldOrder fields) =
-  Fix $ Lam noLoc (PCons noLoc cons args) $ Fix $ Var noLoc v
-  where
-    v = VarName noLoc "$v"
-    args = fmap (\x -> if (field == x) then (PVar noLoc v) else PWildCard noLoc) fields
-
-
-setupUserTypeInfo :: UserTypeCtx -> UserTypeCtx
-setupUserTypeInfo = setupCoreTypeInfo . setupConsInfo . setupUserRecords
-
-setupCoreTypeInfo :: UserTypeCtx -> UserTypeCtx
-setupCoreTypeInfo t = t { userTypeCtx'core = userCoreTypeMap $ M.mapKeys varName'name $ userTypeCtx'types t }
-
-setupConsInfo :: UserTypeCtx -> UserTypeCtx
-setupConsInfo ctx = ctx { userTypeCtx'constrs = getConsInfoMap $ userTypeCtx'types ctx }
-  where
-    getConsInfoMap  :: Map VarName UserType -> Map ConsName ConsInfo
-    getConsInfoMap = foldMap getConsInfo . M.elems
-
-    getConsInfo :: UserType -> Map ConsName ConsInfo
-    getConsInfo ut@UserType{..} = M.fromList $ fmap (toInfo ut) $ zip [0..] $ M.toList userType'cases
-
-    toInfo :: UserType -> (Int, (ConsName, ConsDef)) -> (ConsName, ConsInfo)
-    toInfo userT (tagId, (name, def)) = (name, info)
-      where
-        ty = userType'name userT
-        tyArgs = userType'args userT
-
-        info = ConsInfo
-                { consInfo'tagId   = tagId
-                , consInfo'arity   = arity
-                , consInfo'type    = consTy
-                , consInfo'def     = userT
-                }
-
-        arity = V.length consArgs
-
-        consTy = getConsType (toResultType ty tyArgs) consArgs
-
-        consArgs = getConsTypes def
-
-    getConsType :: Type -> Vector Type -> Type
-    getConsType rhs args = V.foldr (\arg r -> H.arrowT noLoc arg r) rhs args
-
-    toResultType :: VarName -> [VarName] -> Type
-    toResultType name args = H.conT noLoc (varName'name name) (fmap (H.varT noLoc . varName'name) args)
-
-getConsDefArity :: ConsDef -> Int
-getConsDefArity = V.length . getConsTypes
-
-setupUserRecords :: UserTypeCtx -> UserTypeCtx
-setupUserRecords = setupRecFields . setupRecConstrs
-
--- | Fills record field order from user defined types.
-setupRecConstrs :: UserTypeCtx -> UserTypeCtx
-setupRecConstrs ctx = ctx { userTypeCtx'recConstrs = recConstrs }
-  where
-    recConstrs = M.fromList $ mapMaybe getConstr . M.toList . userType'cases =<< types
-
-    getConstr (cons, def) = case def of
-      ConsDef _ -> Nothing
-      RecordCons fields -> Just $ (cons, RecordFieldOrder $
-          fmap (varName'name . recordField'name) $ V.toList fields)
-
-    types = M.elems $ userTypeCtx'types ctx
-
--- | Fills record fields to complete set of fields map for all record fields.
-setupRecFields :: UserTypeCtx -> UserTypeCtx
-setupRecFields ctx = ctx { userTypeCtx'recFields = recFields }
-  where
-    getFieldMap cons fieldOrder@(RecordFieldOrder fields) = fmap (\field -> (field, (cons, fieldOrder))) fields
-
-    recFields = M.fromList $ uncurry getFieldMap =<< (M.toList $ userTypeCtx'recConstrs ctx)
-
-
--- | User-defined type
-data UserType = UserType
-  { userType'name       :: !VarName                -- ^ Type name
-  , userType'args       :: ![VarName]              -- ^ type arguments
-  , userType'cases      :: !(Map ConsName ConsDef) -- ^ List of constructors
-  } deriving (Show, Eq, Data, Typeable)
-
--- | Core representation of user data-types with constructors.
-newtype CoreUserType = CoreUserType
-  { coreUserType'cases  :: Map ConsName CoreConsDef
-  } deriving (Show, Eq, Data)
-
--- | Every user constructor is transformed
--- to combination of tuple and sum type constructors
-data CoreConsDef = CoreConsDef
-  { coreConsDef'sum   :: Maybe (Int, Vector (H.Type () Name)) -- ^ sum type constructor spec
-                                                              --  sometimes we do not need sum nad tuple is enough
-                                                              --  then it's omited (value Nothing)
-  , coreConsDef'tuple :: Vector (H.Type () Name)              -- ^ Multiple arguments are enocded as tuples
-  } deriving (Show, Eq, Data)
-
-
-getConsTypes :: ConsDef -> Vector Type
-getConsTypes = \case
-  ConsDef ts        -> ts
-  RecordCons fields -> fmap recordField'type fields
-
--- | Constructor definition.
-data ConsDef
-  = ConsDef (Vector Type)            -- ^ Simple constructor with collection of type-arguments
-  | RecordCons (Vector RecordField)  -- ^ Record-constructor with named fields
-  deriving (Show, Eq, Data, Typeable)
-
--- | Record named field.
-data RecordField = RecordField
-  { recordField'name :: VarName   -- ^ Name of the field
-  , recordField'type :: Type      -- ^ Type of the field
-  } deriving (Show, Eq, Data, Typeable)
-
--- | Data for low-level rendering of type constructor
--- We need to know it's type, arity and integer tag that is unique within
--- its group of constructor for a given type (global uniqueness is not needed)
-data ConsInfo = ConsInfo
-  { consInfo'type    :: !Type        -- ^ type of the constructor as a function
-  , consInfo'tagId   :: !Int         -- ^ unique integer identifier (within the type scope)
-  , consInfo'arity   :: !Int         -- ^ arity of constructor
-  , consInfo'def     :: !UserType    -- ^ definition where constructor is defined
-  } deriving (Show, Eq, Data, Typeable)
-
--- | Order of names in the record constructor.
--- For constructor
---
--- > User { name :: Text, age :: Int }
---
--- It's going to be
---
--- ["name", "age"]
-newtype RecordFieldOrder = RecordFieldOrder
-  { unRecordFieldOrder :: [Text]
-  } deriving (Show, Eq, Data, Typeable)
-
--- | Type for expression of our language that has type.
---
--- This is phantom type for covenience of type-checker.
-newtype Expr a = Expr Lang
 
 -- | Name of the variable.
 data VarName = VarName
@@ -356,46 +139,11 @@ data Pat
 instance IsString Pat where
   fromString = PVar noLoc . fromString
 
--- | The type represents modules.
-data Module = Module
-  { module'loc       :: !Loc          -- ^ source code location
-  , module'userTypes :: !UserTypeCtx  -- ^ user-defined types
-  , module'binds     :: !(Binds Lang) -- ^ values (functions)
-  } deriving (Data, Typeable)
-
 -- | Type context for inference algorithm
 type TypeContext = H.Context Loc Text
 
 emptyTypeContext :: TypeContext
 emptyTypeContext = H.Context mempty
-
--- | Context for execution (reduction) of expressions of the language
-newtype ExecCtx = ExecCtx
-  { execCtx'vars  :: Map VarName Lang  -- ^ bindings for free variables, outer scope of the execution
-  } deriving newtype (Semigroup, Monoid)
-
--- | Type-inference context.
-data InferCtx = InferCtx
-  { inferCtx'binds :: TypeContext  -- ^ Already derived signatures for
-                                   -- all free variables in the expression
-  , inferCtx'types :: UserTypeCtx  -- ^ User-defined types
-  }
-  deriving stock (Generic)
-  deriving (Semigroup, Monoid) via GenericSemigroupMonoid InferCtx
-
-userTypesToInferCtx :: UserTypeCtx -> InferCtx
-userTypesToInferCtx ctx = InferCtx emptyTypeContext ctx
-
--- | Evaluated module
-data ModuleCtx = ModuleCtx
-  { moduleCtx'types  :: !InferCtx
-  , moduleCtx'exprs  :: !ExecCtx
-  }
-  deriving stock (Generic)
-  deriving (Semigroup, Monoid) via GenericSemigroupMonoid ModuleCtx
-
-getModuleCtxNames :: ModuleCtx -> [Text]
-getModuleCtxNames = M.keys . H.unContext . inferCtx'binds . moduleCtx'types
 
 -- | Alternatives for declarations right-hand-sides.
 -- Because of pattern matching we can have several alternatives
@@ -447,7 +195,7 @@ bindNamesLhs = \case
   FunBind v _ -> Set.singleton v
   PatBind p _ -> freeVarsPat p
 
--- | Main tpye for expressions
+-- | Main tpye for expressions.
 -- It's defined in fix-point style (See package data-fix).
 type Lang = Fix E
 
@@ -528,9 +276,6 @@ data CaseExpr a
 secretVar :: Text -> Text
 secretVar = flip mappend "___"
 
--- | Type tag for type-safe construction
-data SigmaBool
-
 -- | Result of the script can be boolean constant or sigma-expression
 -- that user have to prove.
 data BoolExprResult
@@ -547,71 +292,6 @@ instance FromJSON BoolExprResult where
   parseJSON = withObject "BoolExprResult" $ \obj ->
         (ConstBool <$> obj .: "bool")
     <|> (SigmaResult <$> obj .: "sigma")
-
----------------------------------
--- type constants
-
-intT, boolT, boxT, scriptT, textT, sigmaT, bytesT, unitT :: (IsString v, H.DefLoc loc) => H.Type loc v
-intT    = intT'    H.defLoc
-boolT   = boolT'   H.defLoc
-bytesT  = bytesT'  H.defLoc
-boxT    = boxT'    H.defLoc
-scriptT = scriptT' H.defLoc
-textT   = textT'   H.defLoc
-sigmaT  = sigmaT'  H.defLoc
-unitT   = unitT'   H.defLoc
-
-tupleT :: H.DefLoc loc => [H.Type loc v] -> H.Type loc v
-tupleT = tupleT' H.defLoc
-
-sumT :: IsString v => H.DefLoc loc => [H.Type loc v] -> H.Type loc v
-sumT = sumT' H.defLoc
-
-listT :: H.DefLoc loc => H.Type loc v -> H.Type loc v
-listT = listT' H.defLoc
-
-maybeT :: IsString v => H.DefLoc loc => H.Type loc v -> H.Type loc v
-maybeT = maybeT' H.defLoc
-
-arrowT :: H.DefLoc loc => H.Type loc v -> H.Type loc v -> H.Type loc v
-arrowT = H.arrowT H.defLoc
-
-varT :: H.DefLoc loc => v -> H.Type loc v
-varT = H.varT H.defLoc
-
-funT :: H.DefLoc loc => [H.Type loc v] -> H.Type loc v -> H.Type loc v
-funT args resT = foldr arrowT resT args
-
-argsT :: (IsString v, H.DefLoc loc) => H.Type loc v
-argsT = typeCoreToType argsTuple
-
-constType :: v -> loc -> H.Type loc v
-constType name loc = H.conT loc name []
-
-intT', boolT', boxT', scriptT', textT', sigmaT', bytesT', unitT' :: (IsString v, H.DefLoc loc) => loc -> H.Type loc v
-boxT'    = constType "Box"
-textT'   = constType "Text"
-bytesT'  = constType "Bytes"
-intT'    = constType "Int"
-boolT'   = constType "Bool"
-sigmaT'  = constType "Sigma"
-scriptT' = constType "Script"
-unitT'   = constType "()"
-
-listT' :: loc -> H.Type loc v -> H.Type loc v
-listT' loc a = H.listT loc a
-
-maybeT' :: IsString v => loc -> H.Type loc v -> H.Type loc v
-maybeT' loc a = H.conT loc "Maybe" [a]
-
-tupleT' :: loc -> [H.Type loc v] -> H.Type loc v
-tupleT' loc ts = H.tupleT loc ts
-
-sumT' :: IsString v => loc -> [H.Type loc v] -> H.Type loc v
-sumT' loc ts = H.conT loc "Sum" ts
-
-arrowT' :: loc -> H.Type loc v -> H.Type loc v -> H.Type loc v
-arrowT' = H.arrowT
 
 --------------------------------
 -- instances
@@ -684,7 +364,6 @@ instance H.HasLoc a => H.HasLoc (Rhs a) where
 instance  H.HasLoc (Bind a) where
   type Loc (Bind a) = Loc
   getLoc = H.getLoc . bind'name
-
 
 -------------------------------------------------------------------
 
@@ -785,146 +464,6 @@ sortBinds b = b { binds'decls = onDecls $ binds'decls b }
        FunBind{..} -> pure $ (s, bind'name, Set.toList $ foldMap (freeVarsAlt . fmap freeVars) bind'alts)
        PatBind{..} -> fmap (\name -> (s, name, Set.toList $ (freeVarsAlt . fmap freeVars) bind'alt )) $ Set.toList $ freeVarsPat bind'pat
 
-typeCoreToType :: (H.DefLoc loc, IsString v) => TypeCore -> H.Type loc v
-typeCoreToType = \case
-  IntT      -> intT
-  BoolT     -> boolT
-  BytesT    -> bytesT
-  TextT     -> textT
-  SigmaT    -> sigmaT
-  BoxT      -> boxT
-  a :-> b   -> (arrowT `on` typeCoreToType) a b
-  ListT a   -> listT (typeCoreToType a)
-  TupleT xs -> tupleT $ typeCoreToType <$> xs
-  UnitT     -> unitT
-  MaybeT a  -> maybeT (typeCoreToType a)
-  SumT ts   -> sumT (fmap typeCoreToType ts)
-
-----------------------------------------------------------------
--- Names
-----------------------------------------------------------------
-
--- | Name of monomorphic primop which is used in high level language
-monoPrimopName :: PrimOp a -> Maybe Name
-monoPrimopName = \case
-  OpAdd         -> Just "+"
-  OpSub         -> Just "-"
-  OpMul         -> Just "*"
-  OpDiv         -> Just "/"
-  OpNeg         -> Just "negate"
-  --
-  OpBoolAnd     -> Just "&&"
-  OpBoolOr      -> Just "||"
-  OpBoolNot     -> Just "not"
-  --
-  OpSigAnd       -> Just Const.sigmaAnd
-  OpSigOr        -> Just Const.sigmaOr
-  OpSigPK        -> Just "pk"
-  OpSigDTuple    -> Just "dtuple"
-  OpSigBool      -> Just "toSigma"
-  OpSigListAnd   -> Just "andSigma"
-  OpSigListOr    -> Just "orSigma"
-  OpSigListAll _ -> Nothing
-  OpSigListAny _ -> Nothing
-  --
-  OpCheckSig      -> Just Const.checkSig
-  OpCheckMultiSig -> Just Const.checkMultiSig
-  --
-  OpSHA256      -> Just Const.sha256
-  OpTextLength  -> Just Const.lengthText
-  OpBytesLength -> Just Const.lengthBytes
-  OpTextAppend  -> Just Const.appendText
-  OpBytesAppend -> Just Const.appendBytes
-  OpToBytes   _ -> Nothing
-  OpFromBytes _ -> Nothing
-  --
-  OpArgs _       -> Nothing
-  OpGetBoxId     -> Just Const.getBoxId
-  OpGetBoxScript -> Just Const.getBoxScript
-  OpGetBoxValue  -> Just Const.getBoxValue
-  OpGetBoxArgs _ -> Nothing
-  OpGetBoxPostHeight -> Just $ Const.getBoxPostHeight
-  --
-  OpEnvGetHeight  -> Just Const.getHeight
-  OpEnvGetSelf    -> Just Const.getSelf
-  OpEnvGetInputs  -> Just Const.getInputs
-  OpEnvGetOutputs -> Just Const.getOutputs
-  OpEnvGetDataInputs -> Just Const.getDataInputs
-  -- Polymorphic functions
-  OpShow _ -> Nothing
-  OpEQ _   -> Nothing
-  OpNE _   -> Nothing
-  OpGT _   -> Nothing
-  OpGE _   -> Nothing
-  OpLT _   -> Nothing
-  OpLE _   -> Nothing
-  --
-  OpListMap{}    -> Nothing
-  OpListAt{}     -> Nothing
-  OpListAppend{} -> Nothing
-  OpListLength{} -> Nothing
-  OpListFoldr{}  -> Nothing
-  OpListFoldl{}  -> Nothing
-  OpListFilter{} -> Nothing
-  OpListSum      -> Just "sum"
-  OpListProduct  -> Just "product"
-  OpListAnd      -> Just "and"
-  OpListOr       -> Just "or"
-  OpListAll{}    -> Nothing
-  OpListAny{}    -> Nothing
-
-polyPrimOpName :: PrimOp a -> Maybe Name
-polyPrimOpName = \case
-  OpShow _ -> Just "show"
-  OpEQ _   -> Just "=="
-  OpNE _   -> Just "/="
-  OpGT _   -> Just ">"
-  OpGE _   -> Just ">="
-  OpLT _   -> Just "<"
-  OpLE _   -> Just "<="
-  --
-  OpFromBytes _ -> Just Const.deserialiseBytes
-  OpToBytes   _ -> Just Const.serialiseBytes
-  --
-  OpArgs _       -> Just $ Const.getArgs
-  OpGetBoxArgs _ -> Just $ Const.getBoxArgs
-  --
-  OpListMap{}    -> Just "map"
-  OpListAt{}     -> Just "listAt"
-  OpListAppend{} -> Just "++"
-  OpListLength{} -> Just "length"
-  OpListFoldr{}  -> Just "foldr"
-  OpListFoldl{}  -> Just "foldl"
-  OpListFilter{} -> Just "filter"
-  OpListAll{}    -> Just "all"
-  OpListAny{}    -> Just "any"
-  _              -> Nothing
-
-
--- | List of all monomorphic primops
-monomorphicPrimops :: [PrimOp a]
-monomorphicPrimops =
-  [ OpAdd, OpSub, OpMul, OpDiv, OpNeg
-  , OpBoolAnd, OpBoolOr, OpBoolNot
-  , OpSigAnd, OpSigOr, OpSigPK, OpSigDTuple, OpSigBool, OpSigListAnd, OpSigListOr
-  , OpCheckSig, OpCheckMultiSig
-  , OpSHA256, OpTextLength, OpBytesLength, OpTextAppend, OpBytesAppend
-  , OpEnvGetHeight, OpEnvGetSelf, OpEnvGetInputs, OpEnvGetOutputs, OpEnvGetDataInputs
-  , OpGetBoxId, OpGetBoxScript, OpGetBoxValue, OpGetBoxPostHeight
-  , OpListSum
-  , OpListProduct
-  , OpListAnd
-  , OpListOr
-  ]
-
--- | Name map for substitution of monomorphic primops
-monoPrimopNameMap :: M.Map Name (PrimOp a)
-monoPrimopNameMap = M.fromList
-  [ (nm,op) | op      <- monomorphicPrimops
-            , Just nm <- [ monoPrimopName op ]
-            ]
-
-
 -------------------------------------------------------------------
 
 class ToLang a where
@@ -995,14 +534,4 @@ $(deriveShow1 ''CaseExpr)
 $(deriveEq1   ''E)
 $(deriveShow1 ''E)
 
-deriving newtype instance Show ExecCtx
-deriving newtype instance Eq   ExecCtx
-deriving stock   instance Show ModuleCtx
-deriving stock   instance Eq   ModuleCtx
-deriving stock   instance Show UserTypeCtx
-deriving stock   instance Eq   UserTypeCtx
-deriving stock   instance Show InferCtx
-deriving stock   instance Eq   InferCtx
-deriving stock   instance Show Module
-deriving stock   instance Show (Expr a)
-deriving stock   instance Eq   (Expr a)
+
